style: 레거시 코드, 파일 제거
